backend/requirements.txt
Django==5.0.1
psycopg2-binary==2.9.9
djangorestframework==3.14.0
django-cors-headers==4.3.1
pyjwt ==1.7.1
djangorestframework-jwt == 1.11.0
stripe == 5.0.0
django-environ == 0.11.2
django-filter == 23.5
django-debug-toolbar == 4.3.0
